requirements.txt
fake-bpy-module==20231215
